Documentation suggested enhancements to improve admin user experience
I have added a section for PowerShell prerequisites which helps resolve a number of errors experienced out of the box. This included installing the Azure Az module and ensures the correct Azure subscription is selected for deployment. (lines 24-45)

Further to this, I have tweaked the section (lines 55 to 70) to include the PowerShell cmdlet to unblock files. As its quite common for files to be blocked when downloading from the GitHub repo. I have also changed the order of the copy path line as the previous location was not best placed. 

I hope this helps.