Update rotation script and Azure Durable functions (#242)
This PR

Fix rotation script. While we are in the process of moving away from connection strings, we still need them while Azure Files implement support using managed identity. The rotation script is outdated as az keyvault storage is deprecated so the command doesn't exist in the az cli of the agents. This cmd was useful as it gives you which is the active key to be used for the connection string. Instead we just figure out which one is the oldest one set it in the key vault.
Add new OIDs for service principals
Update nuget package for Azure Durable functions.